+++ client_code/utils/_deprecated.py
@@ -7,7 +7,7 @@
 
 from functools import wraps
 
-__version__ = "2.4.0"
+__version__ = "2.6.2"
 
 
 def deprecated(msg=""):