+++ client_code/uuid.py
@@ -8,9 +8,9 @@
 import anvil.js
 from anvil.js import window as _W
 
-from utils._deprecated import deprecated
+from .utils._deprecated import deprecated
 
-__version__ = "2.4.0"
+__version__ = "2.6.2"
 
 try:
     _js_uuid = _W.uuid