+++ client_code/zod/helpers/dict_util.py
@@ -5,7 +5,7 @@
 #
 # This software is published at https://github.com/anvilistas/anvil-extras
 
-__version__ = "2.4.0"
+__version__ = "2.6.2"
 
 
 def merge_shapes(a, b):
@@ -37,7 +37,7 @@ class DictLike:
 
     def __repr__(self):
         items = self.__dict__.items()
-        return f"{type(self).__name__}({', '.join(f'{k}={v!r}' for k,v in items)})"
+        return f"{type(self).__name__}({', '.join(f'{k}={v!r}' for k, v in items)})"
 
     def __str__(self):
         return str(self.__dict__)