+++ server_code/authorisation.py
@@ -5,10 +5,32 @@
 #
 # This software is published at https://github.com/anvilistas/anvil-extras
 import functools
+from operator import itemgetter
 
 import anvil.users
+from anvil.tables import app_tables
 
-__version__ = "2.4.0"
+__version__ = "2.6.2"
+
+config = {"get_roles": itemgetter("roles")}
+
+
+def set_config(**kwargs):
+    if "get_roles" in kwargs:
+        _set_user_roles_getter(kwargs["get_roles"])
+
+
+def _get_roles_from_table(table_name, user):
+    return getattr(app_tables, table_name).get(user=user)["roles"]
+
+
+def _set_user_roles_getter(option):
+    if option is None:
+        config["get_roles"] = itemgetter("roles")
+    elif isinstance(option, str):  # table name
+        config["get_roles"] = functools.partial(_get_roles_from_table, option)
+    else:
+        raise TypeError("get_roles: option is not valid.")
 
 
 def authentication_required(func):
@@ -24,34 +46,48 @@ def authentication_required(func):
     return wrapper
 
 
+def has_permission(permissions):
+    """Returns True/False depending on whether a user has permission or not"""
+    user = anvil.users.get_user()
+    if user is None:
+        return False
+
+    if isinstance(permissions, str):
+        required_permissions = set([permissions])
+    else:
+        required_permissions = set(permissions)
+
+    try:
+        user_permissions = set(
+            permission["name"]
+            for role in config["get_roles"](user)
+            for permission in role["permissions"]
+        )
+    except TypeError:
+        return False
+
+    return required_permissions.issubset(user_permissions)
+
+
+def check_permissions(permissions):
+    """Checks a users permissions, raises ValueError if user does not have permissions"""
+    if has_permission(permissions):
+        return
+
+    user = anvil.users.get_user()
+    fail = "Authentication" if user is None else "Authorisation"
+
+    raise ValueError(f"{fail} required")
+
+
 def authorisation_required(permissions):
     """A decorator to ensure a user has sufficient permissions to call a server function"""
 
     def decorator(func):
         @functools.wraps(func)
         def wrapper(*args, **kwargs):
-            user = anvil.users.get_user()
-            if user is None:
-                raise ValueError("Authentication required")
-            if isinstance(permissions, str):
-                required_permissions = set([permissions])
-            else:
-                required_permissions = set(permissions)
-            try:
-                user_permissions = set(
-                    [
-                        permission["name"]
-                        for role in user["roles"]
-                        for permission in role["permissions"]
-                    ]
-                )
-            except TypeError:
-                raise ValueError("Authorisation required")
-
-            if not required_permissions.issubset(user_permissions):
-                raise ValueError("Authorisation required")
-            else:
-                return func(*args, **kwargs)
+            check_permissions(permissions)
+            return func(*args, **kwargs)
 
         return wrapper